# routes/api/v1/credential-offer.js
/**
 * IDChain Agent REST API Routes
 * Credential Offer Routes
 */
'use strict';

const router = require('express').Router();
const controller = require('../../../controllers/credential');
const wrap = require('../../../util/asyncwrap').wrap;
const APIResult = require('../../../util/api-result');

router
    .route('/')
    .get(
        wrap(async (req, res, next) => {
            const data = await controller.offer.list(req.wallet, req.query);
            res.locals.result = APIResult.success(data);
            next();
        })
    )
    .post(
        wrap(async (req, res, next) => {
            const data = await controller.offer.create(
                req.wallet,
                req.body.comment,
                req.body.recipientDid,
                req.body.credDefId,
                req.body.credentialLocation
            );
            res.locals.result = APIResult.created(data);
            next();
        })
    );

router.route('/:credentialOfferId').get(
    wrap(async (req, res, next) => {
        const data = await controller.offer.retrieve(req.wallet, req.params.credentialOfferId);
        res.locals.result = data ? APIResult.success(data) : APIResult.notFound();
        next();
    })
);

module.exports = router;