# routes/api/v1/swagger.yaml
swagger: '2.0'
info:
  title: 'DIMS Agent API'
  description: 'DIMS Agent API'
  version: '1.0.0'
schemes:
  - http
basePath: /api/v1
securityDefinitions:
  Bearer:
    description: |
      JWT Authorization header using the Bearer scheme.
      Example: 'Authorization: Bearer {token}'

      Please add Bearer in front of the JWT token:
      > Bearer xxxxx.yyyyy.zzzzz
    type: apiKey
    name: Authorization
    in: header
consumes:
  - application/json
produces:
  - application/json
security:
  - Bearer: []
...
parameters:
...
  walletParam:
    name: wallet
    in: path
    description: |-
      A unique id value identifying a wallet.
      "default" may be used as shorthand for the current user's default wallet.
    type: string
    required: true
  walletHeader:
    name: wallet
    in: header
    description: A unique id value identifying a wallet.
    type: string
  theirDidParam:
    name: theirDid
    in: path
    description: 'Their pairwise connection did'
    type: string
    required: true
  credentialParam:
    name: credential
    in: path
    description: 'A unique id value identifying a credential'
    type: string
    required: true
...
definitions:
  walletId:
    description: '(Optional) A unique id value identifying a wallet'
    type: string
  id:
    description: 'A uniquely identifying string value'
    type: string
  createdAt:
    description: 'Creation Time'
    type: string
  updatedAt:
    description: 'Last update time'
    type: string
  messageId:
    description: 'Sent message id, could be a nonce depending on message format'
    type: string
  messageType:
    description: 'Message Type'
    type: string
  messageObject:
    description: 'Message Object'
    type: object
  senderDid:
    description: 'Sender Did'
    type: string
  recipientDid:
    description: 'Recipient Did'
    type: string
  expireAt:
    description: 'Optional expire time'
    type: string
  metaObject:
    description: 'Additional Metadata'
    type: object
...
  credential_post:
    type: object
    properties:
      wallet:
        $ref: '#/definitions/walletId'
      comment:
        description: '(optional) comment describing the credential'
        type: string
      credentialRequestId:
        description: 'Credential Request _id as stored in API'
        type: string
      values:
        description:
          'Object containing attributes defined in schema as key-value pairs (e.g. {"attrName":"attrValue", "attrName1":"attrValue1"})'
        type: object
    required:
      - credentialRequestId
      - values
...
  credentialoffer_post:
    type: object
    properties:
      wallet:
        $ref: '#/definitions/walletId'
      comment:
        description: '(optional) comment describing the credential offer'
        type: string
      recipientDid:
        description: 'DID for whom to create a credential offer'
        type: string
      credDefId:
        description: 'Credential definition id'
        type: string
      credentialLocation:
        description:
          '(Optional) URL at which key-value object containing credential values can be fetched using HTTP GET'
        type: string
    required:
      - recipientDid
      - credDefId
...
  proof_post:
    type: object
    properties:
      wallet:
        $ref: '#/definitions/walletId'
      proofRequestId:
        description: 'proof request _id'
        type: string
      comment:
        description: '(optional) comment describing the proof'
        type: string
      values:
        description: (Optional) Object containing self-attested-attributes as key-value pairs
        type: object
        example:
          phone: '00001111'
    required:
      - proofRequestId
  indyschema_post:
    type: object
    properties:
      wallet:
        $ref: '#/definitions/walletId'
      name:
        description: 'A unique name of the schema'
        type: string
      version:
        description: 'version for schema'
        type: string
      attrNames:
        description: 'list of attribute names put into the schema'
        type: array
        items:
          type: string
    required:
      - name
      - version
      - attrNames
  schema_post:
    type: object
    required:
      - attributes
      - name
      - version
    properties:
      name:
        type: string
        description: 'A unique name for the schema'
      version:
        type: string
        description: 'Schema version'
      parentSchemaId:
        type: string
        description: 'Schema Id of the parent'
      createCredentialDefinition:
        type: boolean
        default: true
      isRevocable:
        type: boolean
        description:
          'Flag to define if credentials issued with this schema could be revoked. If true it will generate one or multiple Revocation Registries.'
        default: false
      attributes:
        type: array
        description: 'List of attribute names put into the schema'
        items:
          required:
            - name
            - type
          properties:
            name:
              type: string
              description: 'Name of the attribute (Key)'
            type:
              type: string
              description: 'A valid type for attributes, as requested using GET /types in schema compiler API'
    example:
      name: 'Passport'
      version: '1.0'
      isRevocable: false
      parentSchemaId: 'Th7MpTaRZVRYnPiabds81Y:2:ID_NL:2.0'
      attributes:
        - name: 'Given Name'
          type: string
  schema_details:
    allOf:
      - $ref: '#/definitions/schema_post'
      - type: object
        properties:
          isDeprecated:
            type: boolean
          lowLevelSchema:
            type: object
            properties: {}
          credentialDefinitionId:
            type: string
          revocationRegistryId:
            type: string
        example:
          isDeprecated: false
          lowLevelSchema: {}
          credentialDefinitionId: 'Th7MpTaRZVRYnPiabds81Y:3:CL:15:TAG1'
          revocationRegistryId: ''
...
  error_message:
    type: object
    properties:
      message:
        description: 'Informational Message'
        type: string
  connection:
    type: object
    properties:
      my_did:
        description: 'my did'
        type: string
      their_did:
        description: 'their did'
        type: string
      acknowledged:
        description: 'status'
        type: boolean
...
  credential_offer:
    type: object
    properties:
      id:
        $ref: '#/definitions/id'
      createdAt:
        $ref: '#/definitions/createdAt'
      updatedAt:
        $ref: '#/definitions/updatedAt'
      expireAt:
        $ref: '#/definitions/expireAt'
      messageId:
        $ref: '#/definitions/messageId'
      type:
        $ref: '#/definitions/messageType'
      senderDid:
        $ref: '#/definitions/senderDid'
      recipientDid:
        $ref: '#/definitions/recipientDid'
      message:
        $ref: '#/definitions/messageObject'
      meta:
        $ref: '#/definitions/metaObject'
  indy_schema_db:
    description: 'Indy low-level schema as stored in DB'
    type: object
    properties:
      name:
        description: 'Schema Name'
        type: string
      createdAt:
        $ref: '#/definitions/createdAt'
      updatedAt:
        $ref: '#/definitions/updatedAt'
      attrNames:
        type: array
        items:
          description: 'Schema attribute name'
          type: string
      version:
        description: 'Schema Version'
        type: string
      wallet:
        description: 'Owner Wallet Id'
        type: string
      schemaId:
        description: 'Schema Id'
        type: string
      data:
        $ref: '#/definitions/indy_schema'
  indy_schema:
    description: 'Schema as created by indy-sdk and stored on the ledger'
    type: object
    properties:
      id:
        description: 'Schema Id'
        type: string
      attrNames:
        type: array
        items:
          description: 'Schema attribute name'
          type: string
      name:
        description: 'Schema Name'
        type: string
      version:
        description: 'Schema Version'
        type: string
      ver:
        description: 'Schema JSON Version'
        type: string
...
  indy_credential:
    description: 'Credential as created by indy-sdk, fields may vary'
    type: object
    properties:
      referent:
        description: 'Credential Id as stored in wallet'
        type: string
      schema_id:
        description: 'Schema Id'
        type: string
      cred_def_id:
        description: 'Credential Definition Id'
        type: string
      rev_reg_def_id:
        description: '(Optional) Revocation Registry Definition Id, if revocable'
        type: string
      values:
        description: 'Credential Values'
        type: object
      signature:
        description: 'Signature'
        type: string
      signature_correctness_proof:
        description: 'Signature Correctness Proof'
        type: string
...
  credential_message:
    type: object
    properties:
      id:
        $ref: '#/definitions/id'
      createdAt:
        $ref: '#/definitions/createdAt'
      updatedAt:
        $ref: '#/definitions/updatedAt'
      expireAt:
        $ref: '#/definitions/expireAt'
      messageId:
        $ref: '#/definitions/messageId'
      type:
        $ref: '#/definitions/messageType'
      senderDid:
        $ref: '#/definitions/senderDid'
      recipientDid:
        $ref: '#/definitions/recipientDid'
      message:
        $ref: '#/definitions/messageObject'
      meta:
        description: 'Additional Metadata'
        type: object
        properties:
          revocRegId:
            description: '(Optional) Revocation Registry Id'
            type: string
          credRevocId:
            description: '(Optional) Credential Revocation Id'
            type: string
          revocRegDelta:
            description: '(Optional) Revocation Registry Delta'
            type: string
...
  proof_message:
    type: object
    properties:
      id:
        $ref: '#/definitions/id'
      createdAt:
        $ref: '#/definitions/createdAt'
      updatedAt:
        $ref: '#/definitions/updatedAt'
      expireAt:
        $ref: '#/definitions/expireAt'
      messageId:
        $ref: '#/definitions/messageId'
      type:
        $ref: '#/definitions/messageType'
      senderDid:
        $ref: '#/definitions/senderDid'
      recipientDid:
        $ref: '#/definitions/recipientDid'
      message:
        $ref: '#/definitions/messageObject'
  proof:
    type: object
    properties:
      wallet:
        description: 'Recipient wallet Id'
        type: string
      did:
        description: 'Proof Sender/Prover Pairwise Did'
        type: string
      proof:
        description: 'Proof as created by indy-sdk and sent by prover'
        type: object
      status:
        description: 'Proof status (pending or received)'
        type: string
        enum:
          - pending
          - received
      meta:
        description: 'Additional Metadata'
        type: object
        properties:
          proofRequest:
            description: 'Proof request object, necessary for further processing'
            type: object
      isValid:
        description: 'Whether proof is valid or not'
        type: boolean
  pairwise:
    type: object
    properties:
      myDid:
        description: 'My pairwise DID'
        type: string
      theirDid:
        description: 'Their pairwise DID'
        type: string
      acknowledged:
        description: 'If connection is in COMPLETED state'
        type: boolean
      metadata:
        $ref: '#/definitions/metaObject'
...
responses:
...
  no_content:
    description: 'Success - but no content'
  bad_request:
    description: 'Bad Request'
    schema:
      $ref: '#/definitions/error_message'
  not_found:
    description: 'Not Found'
    schema:
      $ref: '#/definitions/error_message'
  unauthorized:
    description: 'Unauthorized'
paths:
...
  /wallet/{wallet}/connection/:
    parameters:
      - $ref: '#/parameters/walletParam'
    get:
      summary: 'List connections/pairwises stored in wallet'
      description: 'List connections/pairwises stored in wallet'
      responses:
        '200':
          description: 'Success'
          schema:
            type: array
            items:
              $ref: '#/definitions/pairwise'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - wallet
        - common
        - issuer
        - verifier
        - prover
  /wallet/{wallet}/connection/{theirDid}:
    parameters:
      - $ref: '#/parameters/walletParam'
      - $ref: '#/parameters/theirDidParam'
    get:
      summary: 'Retrieve connection/pairwise stored in wallet'
      description: 'Retrieve connection/pairwise stored in wallet'
      responses:
        '200':
          description: 'Success'
          schema:
            $ref: '#/definitions/pairwise'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - wallet
        - common
        - issuer
        - verifier
        - prover
  /wallet/{wallet}/credential/:
    get:
      summary: 'List credentials stored in wallet'
      description: 'List credentials stored in wallet'
      parameters:
        - $ref: '#/parameters/walletParam'
        - in: query
          name: schema_id
          type: string
          description: '(Optional) schema_id to filter credentials'
        - in: query
          name: schema_issuer_did
          type: string
          description: '(Optional) schema_issuer_did to filter credentials'
        - in: query
          name: schema_name
          type: string
          description: '(Optional) schema_name to filter credentials'
        - in: query
          name: schema_version
          type: string
          description: '(Optional) schema_version to filter credentials'
        - in: query
          name: issuer_did
          type: string
          description: '(Optional) issuer_did to filter credentials'
        - in: query
          name: cred_def_id
          type: string
          description: '(Optional) cred_def_id to filter credentials'
      responses:
        '200':
          description: 'Success'
          schema:
            type: array
            items:
              $ref: '#/definitions/indy_credential'
        '400':
          $ref: '#/responses/bad_request'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - wallet
        - prover
  /wallet/{wallet}/credential/{credential}:
    parameters:
      - $ref: '#/parameters/walletParam'
      - $ref: '#/parameters/credentialParam'
    get:
      summary: 'Retrieve credential stored in wallet'
      description: 'Retrieve credential stored in wallet'
      responses:
        '200':
          description: 'Success'
          schema:
            $ref: '#/definitions/indy_credential'
      tags:
        - wallet
        - prover
  /connection/:
    get:
      summary: 'List connections'
      description: 'List connections'
      parameters:
        - $ref: '#/parameters/walletHeader'
      responses:
        '200':
          description: ''
          schema:
            type: array
            items:
              $ref: '#/definitions/connection'
        '401':
          $ref: '#/responses/unauthorized'
      tags:
        - connection
        - common
        - issuer
        - verifier
        - prover
  /connection/{myDid}:
    parameters:
      - name: myDid
        in: path
        description: 'my did'
        type: string
        required: true
    get:
      summary: 'Retrieve a connection'
      description: 'Retrieve a connection'
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/connection'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - connection
        - common
        - issuer
        - verifier
        - prover
...
  /indyschema:
    get:
      summary: 'List low-level indy schemas'
      description: 'List low-level indy schemas'
      parameters:
        - $ref: '#/parameters/walletHeader'
      responses:
        '200':
          description: 'Success'
          schema:
            type: array
            items:
              $ref: '#/definitions/indy_schema_db'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - indyschema
        - issuer
    post:
      summary: 'Create a low-level indy schema'
      description: 'Create a low-level indy schema'
      parameters:
        - in: body
          name: data
          schema:
            $ref: '#/definitions/indyschema_post'
      responses:
        '201':
          description: 'Created'
          schema:
            $ref: '#/definitions/indy_schema_db'
        '400':
          $ref: '#/responses/bad_request'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - indyschema
        - issuer
  /indyschema/{schemaid}:
    parameters:
      - description: 'A unique string value identifying this schema.'
        in: path
        name: schemaid
        required: true
        type: string
      - $ref: '#/parameters/walletHeader'
    get:
      summary: 'Retrieve a low-level indy schema'
      description: 'Retrieve a low-level indy schema'
      responses:
        '200':
          description: 'Success'
          schema:
            $ref: '#/definitions/indy_schema'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - indyschema
        - issuer
        - verifier
  /schema:
    get:
      tags:
        - schema
        - issuer
      summary: 'List schemas'
      description: 'List schemas from the logged in user'
      parameters:
        - name: onlyActive
          in: query
          description: '(Optional) Flag to filter all non-active schemas'
          required: false
          type: boolean
        - $ref: '#/parameters/walletHeader'
      responses:
        200:
          description: ''
          schema:
            type: array
            example:
              - name: 'Passport'
                version: '1.0'
                createdAt: '2018-10-21T18:27:23.759Z'
                schemaId: 'Th7MpTaRZVRYnPiabds81Y:2:Passport:1.0'
              - name: 'ID_NL'
                version: '2.0'
                createdAt: '2018-09-21T18:27:23.759Z'
                schemaId: 'Th7MpTaRZVRYnPiabds81Y:2:ID_NL:2.0'
              - name: 'Passport'
                version: '2.0'
                createdAt: '2018-08-21T18:27:23.759Z'
                schemaId: 'Th7MpTaRZVRYnPiabds81Y:2:Passport:2.0'
            items:
              properties:
                name:
                  type: 'string'
                  description: 'Name of the schema'
                version:
                  type: 'string'
                  description: 'Version of the schema'
                schemaId:
                  type: 'string'
                  description: 'Schema Id (as defined in Indy)'
                createdDate:
                  type: 'number'
                  description: 'Creation Date of the schema (EPOCH)'
    post:
      tags:
        - schema
        - issuer
      summary: 'Create a Schema'
      description: 'Create a Schema, Credential Definition & Revocation Registry (if revocable) for the logged in user'
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/schema_post'
        - $ref: '#/parameters/walletHeader'

      responses:
        201:
          description: ''
        400:
          description: 'Schema does not pass the schema compiler/typechecker'
        409:
          description: 'Schema with the same name and version already exists'
  /schema/{schemaid}:
    get:
      tags:
        - schema
        - issuer
        - verifier
      summary: 'Retrieve a schema'
      description: 'Retrieve a schema'
      parameters:
        - name: schemaid
          in: path
          description: 'A unique string value identifying this schema.'
          required: true
          type: string
        - $ref: '#/parameters/walletHeader'
      responses:
        200:
          description: 'Schema Details'
          schema:
            $ref: '#/definitions/schema_details'
        404:
          description: 'Schema with that ID not found in database'

    patch:
      tags:
        - schema
        - issuer
      summary: 'Alter schema by the given operation'
      description: 'Alter schema by the given operation'
      parameters:
        - $ref: '#/parameters/walletHeader'
        - name: schemaid
          in: path
          description: 'A unique string value identifying this schema.'
          required: true
          type: string
        - in: body
          name: body
          required: false
          schema:
            type: 'object'
            required:
              - 'operation'
            properties:
              operation:
                type: 'string'
                enum:
                  - 'revoke'
      responses:
        204:
          description: 'Successfully altered schema'
        422:
          description: 'Schema does not support revocations, or is already revoked'
...
  /credentialoffer:
    get:
      summary: 'List credential offers'
      description: 'List credential offers'
      parameters:
        - $ref: '#/parameters/walletHeader'
      responses:
        '200':
          description: 'Success'
          schema:
            type: array
            items:
              $ref: '#/definitions/credential_offer'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - credential
        - issuer
        - prover
    post:
      summary: 'Create and send a credential offer'
      description: 'Create and send a credential offer'
      parameters:
        - in: body
          name: data
          schema:
            $ref: '#/definitions/credentialoffer_post'
      responses:
        '201':
          description: 'Created'
          schema:
            $ref: '#/definitions/credential_offer'
        '400':
          $ref: '#/responses/bad_request'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - credential
        - issuer
  /credentialoffer/{id}:
    parameters:
      - description: 'A unique string value identifying a credential offer.'
        in: path
        name: id
        required: true
        type: string
      - $ref: '#/parameters/walletHeader'
    get:
      summary: 'Retrieve a credential offer'
      description: 'Retrieve a credential offer'
      responses:
        '200':
          description: 'Success'
          schema:
            type: array
            items:
              $ref: '#/definitions/credential_offer'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - credential
        - issuer
        - prover
    delete:
      summary: 'Delete a credential offer'
      description: 'Delete a credential offer'
      responses:
        '204':
          $ref: '#/responses/no_content'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - credential
        - issuer
        - prover
...
  /credential/:
    get:
      summary: 'List credentials issued by a wallet'
      description: 'List credentials issued by a wallet'
      parameters:
        - $ref: '#/parameters/walletHeader'
        - in: query
          name: recipientDid
          type: string
          description: '(Optional) recipientDid to filter by'
      responses:
        '200':
          description: 'Success'
          schema:
            type: array
            items:
              $ref: '#/definitions/credential_message'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - credential
        - issuer
    post:
      summary: 'Issue/Create a credential and send to recipient'
      description: 'Issue/Create a credential and send to recipient'
      parameters:
        - in: body
          name: data
          schema:
            $ref: '#/definitions/credential_post'
      responses:
        '201':
          description: 'Created'
          schema:
            $ref: '#/definitions/credential_message'
        '400':
          $ref: '#/responses/bad_request'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - credential
        - issuer
  /credential/{id}:
    parameters:
      - description: 'A unique string value identifying this credential.'
        in: path
        name: id
        required: true
        type: string
      - $ref: '#/parameters/walletHeader'
    get:
      summary: 'Retrieve a credential message issued/sent by a wallet'
      description: 'Retrieve a credential message issued/sent by a wallet'
      responses:
        '200':
          description: 'Success'
          schema:
            $ref: '#/definitions/credential_message'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - credential
        - issuer

  # Endpoint is WIP
...
  /proof:
    get:
      summary: 'List received proofs'
      description: 'List received proofs'
      parameters:
        - $ref: '#/parameters/walletHeader'
      responses:
        '200':
          description: 'Success'
          schema:
            type: array
            items:
              $ref: '#/definitions/proof'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - proof
        - verifier
    post:
      summary: 'Create and send a Proof'
      description: 'Create and send a Proof'
      parameters:
        - in: body
          name: data
          schema:
            $ref: '#/definitions/proof_post'
      responses:
        '201':
          description: 'Created'
          schema:
            $ref: '#/definitions/proof_message'
        '400':
          $ref: '#/responses/bad_request'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - proof
        - prover
  /proof/{id}:
    parameters:
      - description: 'A unique string value identifying a proof.'
        in: path
        name: id
        required: true
        type: string
      - $ref: '#/parameters/walletHeader'
    get:
      summary: 'Retrieve a received proof (including whether it is valid)'
      description: 'Retrieve a received proof (including whether it is valid)'
      responses:
        '200':
          description: 'Success'
          schema:
            $ref: '#/definitions/proof'
        '400':
          $ref: '#/responses/bad_request'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - proof
        - verifier
    delete:
      summary: 'Delete a received proof'
      description: 'Delete a received proof'
      responses:
        '204':
          $ref: '#/responses/no_content'
        '401':
          $ref: '#/responses/unauthorized'
        '404':
          $ref: '#/responses/not_found'
      tags:
        - proof
        - verifier
...